Minor changes
git-svn-id: https://op4j.svn.sourceforge.net/svnroot/op4j/trunk/op4j@1076 2d6b427a-3b9d-4b24-b9d1-46c41b8758ae

+++ src/main/java/org/op4j/functions/FnFloat.java
@@ -2829,8 +2829,8 @@ public final class FnFloat {
         
         @Override
         protected Float nullAsNullExecute(final Float input, final ExecCtx ctx) throws Exception {
-            return Float.valueOf(BigDecimal.valueOf(input.floatValue())
-                    .remainder(BigDecimal.valueOf(this.divisor.floatValue())).floatValue());
+            return Float.valueOf(BigDecimal.valueOf(input.doubleValue())
+                    .remainder(BigDecimal.valueOf(this.divisor.doubleValue())).floatValue());
         }   
     }
